# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Order API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "user_id": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "enum": [
            "ordered",
            "confirmed",
            "cancelled"
          ]
        }
      },
      "required": [
        "name",
        "price",
        "user_id",
        "status"
      ]
    }
  },
  "paths": {
    "/users/orders": {
      "get": {
        "summary": "Retrieve orders",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/orders": {
      "post": {
        "summary": "Create an order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve orders for a user",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "delete": {
        "summary": "Delete an order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/cancel": {
      "put": {
        "summary": "Cancel an order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/confirm": {
      "put": {
        "summary": "Confirm an order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orders/ordered": {
      "get": {
        "summary": "Retrieve ordered orders",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/orders/confirmed": {
      "get": {
        "summary": "Retrieve confirmed orders",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/orders/cancelled": {
      "get": {
        "summary": "Retrieve cancelled orders",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    }
  }
}